requirements.txt
setuptools
olga>=1.2.4
jupyter>=1.0
git+https://github.com/uio-bmi/immuneML.git@master
pandas>=1.3.3
numpy>=1.21.2
git+https://github.com/uio-bmi/dagsim.git@main
plotly>=5.3.1
graphviz
PyYAML~=5.4.1
scikit-learn~=1.0
fisher~=0.1.9
statsmodels